+++ auditlog/context.py
@@ -3,7 +3,6 @@ import threading
 import time
 from functools import partial
 
-from django.contrib.auth import get_user_model
 from django.db.models.signals import pre_save
 
 from auditlog.models import LogEntry
@@ -55,12 +54,8 @@ def _set_actor(user, sender, instance, signal_duid, **kwargs):
     else:
         if signal_duid != auditlog["signal_duid"]:
             return
-        auth_user_model = get_user_model()
-        if (
-            sender == LogEntry
-            and isinstance(user, auth_user_model)
-            and instance.actor is None
-        ):
+
+        if sender == LogEntry and instance.actor is None:
             instance.actor = user
 
         instance.remote_addr = auditlog["remote_addr"]