+++ auditlog/admin.py
@@ -26,10 +26,5 @@ class LogEntryAdmin(admin.ModelAdmin, LogEntryAdminMixin):
     ]
 
     def has_add_permission(self, request):
-        return False
-
-    def has_change_permission(self, request, obj=None):
-        return False
-
-    def has_delete_permission(self, request, obj=None):
+        # As audit admin doesn't allow log creation from admin
         return False